move routing configuration to daemon
